Remove require_relative and fix autoload class names
Require_relative was being used in a few places and some class names
didn't match file paths.
