--- app/models/manageiq/providers/lenovo/physical_infra_manager/parsers/parsers.rb
@@ -1,12 +1,0 @@
-module ManageIQ::Providers::Lenovo
-  #
-  # Module that contains structures that provides the parse of LXCA
-  #   informations to MiQ format
-  #
-  module Parsers
-  end
-end
-
-require_relative 'components/physical_server_parser'
-require_relative 'components/config_pattern_parser'
-require_relative 'components/physical_rack_parser'
